Split `new.jl` into topic files and prune duplicated tests
`test/new.jl` dominated the parallel test run (14 minutes on CI, twice the
next longest file). Split it into `add.jl`, `develop.jl`, `update.jl`,
`pkgtest.jl`, `status.jl`, `concurrent.jl`, `repo.jl` and
`project_files.jl` so the runner can spread it over workers.

Speed-ups and pruning:

- `isolate_and_pin_registry` fetches only the pinned registry commit with
  `--depth=1` instead of cloning the full General history (the
  `add: resolve tiers` testset went from ~105s to ~12s locally).
- Remove testsets from `pkg.jl` that were already covered by the `new.jl`
  rewrite (stdlib add/update, resolver error names, nonexisting packages,
  `libgit2`/tarball downloads, `up` without manifest, invalid repo URLs,
  canonicalized manifest paths, `up` pruning), plus internal duplicates
  (`instantiate: input checking` was a copy of the `update` one).
- Move all `X: REPL` lowering testsets into `repl.jl` and turn the heavy
  REPL behavior blocks there into `TEST_MODE` lowering checks; the real
  `Pkg.test` spawns and the `PackageCompiler` clone are gone. Fix the
  `up` lowering checks, which destructured the empty args vector as `opts`.
